+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## v2 - 2025-11-28
+
+Add the `is_av1_video()` function for [detecting AV1-encoded videos](https://alexwlchan.net/2025/detecting-av1-videos/).
+
 ## v1 - 2025-11-28
 
-Initial release on PyPI, to test the release mechanism.
+Initial release. Included functions:
+
+* `date_matches_any_format`
+* `date_matches_format`
+* `find_all_dates`
+* `reformat_date`